+++ foundation_stdlib/init.lua
@@ -1,10 +1,9 @@
-local mod = foundation.new_module("foundation_stdlib", "1.18.0")
+local mod = foundation.new_module("foundation_stdlib", "1.19.0")
 
 mod:require("lib/encoding_tables.lua")
 
 mod:require("lib/table.lua")
 
-mod:require("lib/color.lua")
 mod:require("lib/vector.lua")
 mod:require("lib/cuboid.lua")
 mod:require("lib/direction.lua")
@@ -24,6 +23,7 @@ mod:require("lib/type_conversion.lua")
 mod:require("lib/value.lua")
 mod:require("lib/rect.lua")
 mod:require("lib/symbols.lua")
+mod:require("lib/color.lua")
 
 mod:require("lib/waves.lua")
 mod:require("lib/easers.lua")